Refactor : alert와  confirm 컴포넌트 분리 및 코드 수정
