+++ src/components/Alert.jsx
@@ -1,12 +1,8 @@
 import { useRef, useCallback, useEffect } from "react"
-import { useDispatch } from "react-redux"
-import { remove } from "../redux/cart"
 import styled from "styled-components"
 
-export default function Alert({ type, setState, state, title, itemId }) {
+export default function Alert({ type, setState, state }) {
   const modalRef = useRef()
-
-  const dispatch = useDispatch()
 
   const closeModal = (e) => {
     if (modalRef.current === e.target) {
@@ -20,7 +16,7 @@ export default function Alert({ type, setState, state, title, itemId }) {
         setState(false)
       }
     },
-    [setState, state]
+    [state]
   )
 
   useEffect(() => {
@@ -51,38 +47,6 @@ export default function Alert({ type, setState, state, title, itemId }) {
             </DeleteLabel>
             <PartialContent className="text-lg font-bold">선택한 상품이 없습니다.</PartialContent>
           </PartialWrapper>
-        </div>
-      </>
-    )
-  } else if (type === "partialDelete") {
-    return (
-      <>
-        <input type="checkbox" id="my-modal" checked={isChecked(state)} className="modal-toggle" readOnly />
-        <div className="modal" ref={modalRef} onClick={closeModal}>
-          <ModalContent className="modal-box">
-            <Content className="font-bold text-lg">{title}을 삭제하시겠습니까?</Content>
-            <div className="modal-action">
-              <ConFirm
-                htmlFor="my-modal"
-                className="btn btn-primary"
-                onClick={() => {
-                  setState((modal) => !modal)
-                }}
-              >
-                취소
-              </ConFirm>
-              <button
-                className="btn btn-secondary"
-                onClick={() => {
-                  setState((modal) => !modal)
-                  console.log(itemId)
-                  dispatch(remove({ id: itemId }))
-                }}
-              >
-                확인
-              </button>
-            </div>
-          </ModalContent>
         </div>
       </>
     )
@@ -120,22 +84,4 @@ const DeleteLabel = styled.label`
 const PartialContent = styled.h3`
   font-size: 1.5rem;
   margin: auto;
-`
-
-const ModalContent = styled.div`
-  height: 14rem;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-`
-
-const Content = styled.h3`
-  font-size: 1.5rem;
-  font-weight: 700;
-  text-align: center;
-  margin: 1rem 0;
-`
-
-const ConFirm = styled.label`
-  color: white;
 `
